Use powershell to open report in wsl
