+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/DashboardReporterTests.cs
@@ -6,8 +6,8 @@ using Stryker.Core.Options;
 using Stryker.Core.Options.Inputs;
 using Stryker.Core.ProjectComponents.TestProjects;
 using Stryker.Core.Reporters;
-using Stryker.Core.Reporters.Html.ProcessWrapper;
 using Stryker.Core.Reporters.Json;
+using Stryker.Core.Reporters.WebBrowserOpener;
 using Xunit;
 
 namespace Stryker.Core.UnitTest.Reporters
